Test against package names containing periods
The test rejects any package name containing a period.

Explanation: When the packages are compiled into sublime-package bundles
any filenames containing a period confuse the Sublime Text plugin
loader. This is on account of python (sub)modules being separated by a
period. It attempts to, for example, "importlib.import_module(
'Module.app Menu')". This maps to looking for module "Module" and "app
Menu" within for which the sublime-package would be
Module.sublime-package but is not where our plugin is packaged to by
Package Control.